Fold the vlm-exam detection contract into anthropic_claude@v4
v4 shipped in #2850 this week and is not in any release yet, so the
detection changes land in it directly instead of a new v5 version.
The dropped v5 module was identical except for its identifiers.
