+++ inference/core/workflows/core_steps/loader.py
@@ -482,9 +482,6 @@ from inference.core.workflows.core_steps.models.foundation.anthropic_claude.v3 i
 from inference.core.workflows.core_steps.models.foundation.anthropic_claude.v4 import (
     AnthropicClaudeBlockV4,
 )
-from inference.core.workflows.core_steps.models.foundation.anthropic_claude.v5 import (
-    AnthropicClaudeBlockV5,
-)
 
 if not ENABLE_TENSOR_DATA_REPRESENTATION:
     from inference.core.workflows.core_steps.models.foundation.clip.v1 import (
@@ -1764,7 +1761,6 @@ def load_blocks() -> List[Type[WorkflowBlock]]:
         AnthropicClaudeBlockV2,
         AnthropicClaudeBlockV3,
         AnthropicClaudeBlockV4,
-        AnthropicClaudeBlockV5,
         SpaceXAIBlockV1,
         SpaceXAIBlockV2,
         CosineSimilarityBlockV1,
